refactor: napraw wszystkie bugprone-narrowing-conversions (169/169) (part 31)
Ostatnia duża odłożona kategoria z pełnego skanu clang-tidy. Wszystkie
znaleziska to jeden z dwóch wzorców: qsizetype/size_t zawężane do int
(m.in. .size()/.indexOf()), oraz uint64_t/int64_t zawężane do double/float
w obliczeniach ML (LearningEngine*) i statystykach — w obu przypadkach
po prostu dodano jawny static_cast, bez zmiany zachowania. Check nie ma
auto-fixa (puste Replacements w wyeksportowanym YAML), więc wszystko
zrobione ręcznie w 55 plikach src/core/.

Brak nowych realnych bugów w stylu CanFrameModel dlc>256 — dwie rzeczy
odnotowane jako pre-istniejące (nie zmienione dalej, to zmiana projektowa
nie poprawka konwersji):
- DbcParser.cpp: sygnały 64-bit tracą ~11 bitów precyzji przy rzutowaniu na
  double (scale/offset) — zachowanie sprzed tej zmiany, tylko jawne teraz.
- CanRecorder.cpp: ZSTD_compressBound() (size_t) rzutowane na qsizetype —
  bezpieczne w praktyce (ograniczone rozmiarem wejścia).

1324/1324 testów przechodzi (w tym CanSniffer.Start_DriverOpenFails_EmitsError
naprawiony w poprzedniej turze).

+++ src/core/ArxmlParser.cpp
@@ -10,7 +10,7 @@
 namespace {
 
 QString shortRef(const QString &ref) {
-    int idx = ref.lastIndexOf('/');
+    int idx = static_cast<int>(ref.lastIndexOf('/'));
     return (idx >= 0) ? ref.mid(idx + 1) : ref;
 }
