+++ src/core/CanPlayer.h
@@ -33,8 +33,8 @@ public:
 
     bool isPlaying() const { return m_playing; }
     bool isPaused() const  { return m_loaded && !m_playing; }
-    int totalFrames() const { return m_frames.size(); }
-    int frameCount()  const { return m_frames.size(); }
+    int totalFrames() const { return static_cast<int>(m_frames.size()); }
+    int frameCount()  const { return static_cast<int>(m_frames.size()); }
     int currentFrame() const { return m_currentIndex; }
     CanFrame frameAt(int idx) const { return (idx >= 0 && idx < m_frames.size()) ? m_frames[idx] : CanFrame{}; }
     /// Zwraca override jeśli istnieje dla idx, inaczej oryginał.
